requirements.txt
fastapi
uvicorn
pillow
clip-by-openai
torch==2.0.1+cpu
torchvision==0.15.2+cpu
--index-url https://download.pytorch.org/whl/cpu